added TestUtils.scala
